requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-autocomplete-light==3.3.2
django-jet==1.0.8
django-map-widgets==0.2.0
django-smart-selects==1.5.4
django-solo==1.1.3
docopt==0.4.0
idna==2.8
Pillow==5.3.0
psycopg2-binary==2.7.5
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
